SEMANA 13/mi_proyecto_flask/requirements.txt
Flask
mysql-connector-python
sqlite3
